Moved current blockchain related code into its own blockchain module.
