+++ blockchain/go.mod
@@ -1,0 +1,3 @@
+module github.com/sammyklan3/finality/blockchain
+
+go 1.24.5
