+++ go.mod
@@ -1,3 +1,3 @@
-module finality
+module github.com/sammyklan3/finality
 
 go 1.24.5
